requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
WTForms==3.1.1
Werkzeug==3.0.1
qrcode[pil]==7.4.2
Pillow>=10.0.0
python-dotenv==1.0.0
email-validator==2.3.0